Reworked component structure
