--- instances/src/components/ServiceInstanceDetails/ServiceInstanceToolbar/styled.js
@@ -1,16 +1,0 @@
-import styled from 'styled-components';
-
-export const ServiceInstanceToolbarHeadline = styled.header``;
-
-export const ServiceInstanceToolbarHeadlineLink = styled.a`
-  color: #0b74de;
-  text-decoration: none !important;
-  cursor: pointer;
-`;
-
-export const ServiceClassDescription = styled.div`
-  color: #74777a;
-  font-size: 16px;
-  text-align: left;
-  padding: 0 30px;
-`;
